rename files and objects to a lowercase plural pattern
