+++ docs/src/main/scala/cosas/ops/typeSet/AggregateProperties.scala.md
@@ -1,6 +1,6 @@
 
 ```scala
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
 import ohnosequences.cosas._, AnyFn._, AnyTypeSet._
 ```
