+++ docs/src/main/scala/cosas/ops/typeSet/Mappers.scala.md
@@ -1,6 +1,6 @@
 
 ```scala
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
 import ohnosequences.cosas._, AnyFn._, AnyTypeSet._
 import shapeless._, poly._
