+++ docs/src/main/scala/cosas/ops/typeSet/Reorder.scala.md
@@ -5,7 +5,7 @@ Just a combination of ~:~ and Take (reordering set)
 
 
 ```scala
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
 import ohnosequences.cosas._, AnyFn._, AnyTypeSet._
