+++ docs/src/main/scala/cosas/ops/typeSet/Subtract.scala.md
@@ -1,7 +1,7 @@
 ## Subtract one set from another
 
 ```scala
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
 import ohnosequences.cosas._, AnyFn._, AnyTypeSet._
