+++ docs/src/main/scala/cosas/ops/typeSet/Take.scala.md
@@ -1,7 +1,7 @@
 ## Taking a subset
 
 ```scala
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
 import ohnosequences.cosas._, AnyFn._, AnyTypeSet._
