+++ docs/src/main/scala/cosas/ops/typeSet/Union.scala.md
@@ -1,7 +1,7 @@
 ## Sum/union of two type sets
 
 ```scala
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
 import ohnosequences.cosas._, AnyFn._, AnyTypeSet._
