+++ src/main/scala/cosas/ops/records/Get.scala
@@ -5,10 +5,10 @@ It's like `Lookup`, but it removes the element
 
 */
 
-package ohnosequences.cosas.ops.record
+package ohnosequences.cosas.ops.records
 
-import ohnosequences.cosas._, fn._, denotation._, property._, record._
-import ops.typeSet._
+import ohnosequences.cosas._, fns._, types._, properties._, records._
+import ops.typeSets._
 
 @annotation.implicitNotFound(msg = "Can't get property ${P} of the record ${R}")
 trait Get[R <: AnyRecord, P <: AnyProperty]
