+++ src/main/scala/cosas/ops/typeSets/Mappers.scala
@@ -1,6 +1,6 @@
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
-import ohnosequences.cosas._, fn._, typeSet._
+import ohnosequences.cosas._, fns._, typeSets._
 import shapeless._, poly._
