+++ src/main/scala/cosas/ops/typeSets/Pop.scala
@@ -4,9 +4,9 @@
 Returns the element of the set of the given type and the rest of the set
 */
 
-package ohnosequences.cosas.ops.typeSet
+package ohnosequences.cosas.ops.typeSets
 
-import ohnosequences.cosas._, fn._, typeSet._
+import ohnosequences.cosas._, fns._, typeSets._
 
 @annotation.implicitNotFound(msg = "Can't pop an element of type ${E} from the set ${S}")
 trait Pop[S <: AnyTypeSet, E] extends Fn1[S] {
